commit 2-7

+++ src/main/java/com/vineyarg/demo/entidades/Producto.java
@@ -32,8 +32,8 @@ public class Producto implements Serializable {
     @ManyToOne
     private Productor productor;
     private String sku;
-    private int cantidadVecesValorado;
-    private int cantidadValoraciones;
+    private Double cantidadVecesValorado;
+    private Double cantidadValoraciones;
     private Double promedioValoraciones;
     @ManyToOne
     private Compra compra;
@@ -47,7 +47,7 @@ public class Producto implements Serializable {
     public Producto() {
     }
 
-    public Producto(String id, String nombre, Integer cantidad, Double precio, String descripcion, String varietal, Productor productor, String sku, int cantidadVecesValorado, int cantidadValoraciones, Double promedioValoraciones, Compra compra, boolean alta, Set<Imagenes> imagenes) {
+    public Producto(String id, String nombre, Integer cantidad, Double precio, String descripcion, String varietal, Productor productor, String sku, Double cantidadVecesValorado, Double cantidadValoraciones, Double promedioValoraciones, Compra compra, boolean alta, Set<Imagenes> imagenes) {
         this.id = id;
         this.nombre = nombre;
         this.cantidad = cantidad;
@@ -63,6 +63,8 @@ public class Producto implements Serializable {
         this.alta = alta;
         this.imagenes = imagenes;
     }
+
+   
 
     public Compra getCompra() {
         return compra;
@@ -136,22 +138,23 @@ public class Producto implements Serializable {
         this.sku = sku;
     }
 
-    public int getCantidadVecesValorado() {
+    public Double getCantidadVecesValorado() {
         return cantidadVecesValorado;
     }
 
-    public void setCantidadVecesValorado(int cantidadVecesValorado) {
+    public void setCantidadVecesValorado(Double cantidadVecesValorado) {
         this.cantidadVecesValorado = cantidadVecesValorado;
     }
 
-    public int getCantidadValoraciones() {
+    public Double getCantidadValoraciones() {
         return cantidadValoraciones;
     }
 
-    public void setCantidadValoraciones(int cantidadValoraciones) {
+    public void setCantidadValoraciones(Double cantidadValoraciones) {
         this.cantidadValoraciones = cantidadValoraciones;
     }
 
+    
     public Double getPromedioValoraciones() {
         return promedioValoraciones;
     }
